+++ bin/main.js
@@ -4,8 +4,8 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var cmd = require("commander");
 var utils_1 = require("./lib/utils");
 cmd
-    .version("0.3.7", "-v, --version")
-    .usage("[options] <files ...>")
+    .version("0.6.0", "-v, --version")
+    .usage("[options] [file,...]")
     .option("--run <commands>", "run commands then wait for changes to re-run.")
     .option("--wait-first", "wait first, commands will run on changes.")
     .option("--runp-npm <scripts>", "NPM scripts to run parallel.")